requirements.txt
absl-py==0.9.0
attrs==19.3.0
Automat==0.8.0
blinker==1.4
blis==0.4.1
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloud-init==20.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cycler==0.10.0
cymem==2.0.4
dbus-python==1.2.16
decorator==4.4.2
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.0.0
en-core-web-sm==2.2.0
entrypoints==0.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.10.20
httpcore==0.9.1
httplib2==0.14.0
httpx==0.13.3
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.10.0
joblib==0.17.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kiwisolver==1.3.1
langdetect==1.0.8
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
matplotlib==3.3.3
more-itertools==4.2.0
murmurhash==1.0.4
netifaces==0.10.4
numpy==1.19.4
oauthlib==3.1.0
pandas==1.1.4
Pillow==8.0.1
plac==0.9.6
preshed==3.0.4
psycopg2-binary==2.8.6
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.1
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-Levenshtein==0.12.0
pytz==2020.4
PyYAML==5.3.1
rfc3986==1.4.0
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.5.4
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sklearn==0.0
sniffio==1.2.0
spacy==2.2.0
srsly==1.0.4
ssh-import-id==5.10
systemd-python==234
thinc==7.1.1
threadpoolctl==2.1.0
tqdm==4.52.0
traitlets==4.3.3
Twisted==18.9.0
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
wadllib==1.3.3
wasabi==0.8.0
zipp==1.0.0
zope.interface==4.7.1
